+++ goarmorconfigs/configer.go
@@ -1,9 +1,9 @@
 package goarmorconfigs
 
 type Configer interface {
-	ServerDebuggingLevel() uint64
+	ServerDebuggingLevel() int
 }
 
-func (c *Config) ServerDebuggingLevel() uint64 {
+func (c *Config) ServerDebuggingLevel() int {
 	return c.Server.DebuggingLevel
 }
